requirements-dev.txt
-r requirements-acceptance.txt

qontract-reconcile
mypy
ruff==0.5.1
